Move-Attack-Move and some other functionality
I added the ability to set in a module a GlobalInt flag if people want to allow Move, Attack, Move style like 5th edition D&D. or Spring Attack feat. If you create in the GlobalInt section of a module called rsMoveAttackMove and set it equal to 1 then the new style is available.  The program will keep track of whether a player used their "Action" or not, all the buttons for traits and spells remain available, but if you already used an Action it will give a popup warning, "already used action". I hope in future programming to allow Swift Action traits and the Bonus actions etc. For this to work, you also need to update TraitSelect.cs and SpellSelect.cs because those both require a small amount of code to check whether a player has already used the Action that round.

I also allow multiattacks that a player can control the target of each attack.